PRO-7870: Move case set up to utils
